flat-api/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.4.1
    # via django
django-health-check==3.16.4
    # via -r requirements.in
django==3.2.9
    # via
    #   -r requirements.in
    #   django-health-check
    #   djangorestframework
djangorestframework==3.12.4
    # via -r requirements.in
gitdb==4.0.9
    # via gitpython
gitpython==3.1.26
    # via -r requirements.in
pytz==2021.3
    # via django
smmap==5.0.0
    # via gitdb
sqlparse==0.4.2
    # via django
